requirements.txt
# Core dependencies
playwright>=1.44.0
pydantic>=2.7.0
openai>=1.0.0
python-dotenv>=1.0.0

# Optional performance enhancement
orjson>=3.10.0
